requirements.txt
aiodns==3.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
asttokens==3.0.0
attrs==25.3.0
ccxt==4.4.95
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
cryptography==45.0.5
cycler==0.12.1
dateparser==1.2.2
decorator==5.2.1
dill==0.4.0
et_xmlfile==2.0.0
executing==2.2.0
fonttools==4.59.0
frozenlist==1.7.0
idna==3.10
imageio==2.37.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
jedi==0.19.2
joblib==1.5.1
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
llvmlite==0.44.0
matplotlib==3.10.3
matplotlib-inline==0.1.7
multidict==6.6.3
mypy_extensions==1.1.0
numba==0.61.2
numpy==2.2.6
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
parso==0.8.4
pillow==11.3.0
plotly==5.24.1
prompt_toolkit==3.0.51
propcache==0.3.2
pure_eval==0.2.3
pycares==4.9.0
pycparser==2.22
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
schedule==1.2.2
scikit-learn==1.7.1
scipy==1.16.0
setuptools==80.9.0
six==1.17.0
stack-data==0.6.3
tenacity==9.1.2
threadpoolctl==3.6.0
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
vectorbt==0.28.0
wcwidth==0.2.13
widgetsnbextension==4.0.14
yarl==1.20.1
